Extract scooter_action_button to file
