+++ lib/home_screen.dart
@@ -9,6 +9,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'package:unustasis/control_screen.dart';
 import 'package:unustasis/domain/icomoon.dart';
 import 'package:unustasis/driving_screen.dart';
+import 'package:unustasis/interfaces/phone/scooter_action_button.dart';
 import 'package:unustasis/interfaces/phone/scooter_power_button.dart';
 import 'package:unustasis/onboarding_screen.dart';
 import 'package:unustasis/scooter_service.dart';
@@ -380,55 +381,5 @@ class _HomeScreenState extends State<HomeScreen> {
     } else {
       widget.scooterService.optionalAuth = true;
     }
-  }
-}
-
-class ScooterActionButton extends StatelessWidget {
-  const ScooterActionButton({
-    super.key,
-    required void Function()? onPressed,
-    required IconData icon,
-    Color? iconColor,
-    required String label,
-  })  : _onPressed = onPressed,
-        _icon = icon,
-        _iconColor = iconColor,
-        _label = label;
-
-  final void Function()? _onPressed;
-  final IconData _icon;
-  final String _label;
-  final Color? _iconColor;
-
-  @override
-  Widget build(BuildContext context) {
-    Color mainColor = _iconColor ??
-        (_onPressed == null
-            ? Theme.of(context).colorScheme.onBackground.withOpacity(0.2)
-            : Theme.of(context).colorScheme.onBackground);
-    return Column(
-      children: [
-        OutlinedButton(
-          style: OutlinedButton.styleFrom(
-            padding: const EdgeInsets.all(24),
-            side: BorderSide(
-              color: mainColor,
-            ),
-          ),
-          onPressed: _onPressed,
-          child: Icon(
-            _icon,
-            color: mainColor,
-          ),
-        ),
-        const SizedBox(height: 16),
-        Text(
-          _label,
-          style: TextStyle(
-            color: mainColor,
-          ),
-        ),
-      ],
-    );
   }
 }
