+++ lib/interfaces/phone/scooter_action_button.dart
@@ -1,0 +1,56 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+import 'package:flutter_i18n/flutter_i18n.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+import 'package:unustasis/domain/scooter_state.dart';
+import 'package:unustasis/scooter_service.dart';
+
+class ScooterActionButton extends StatelessWidget {
+  const ScooterActionButton({
+    super.key,
+    required void Function()? onPressed,
+    required IconData icon,
+    Color? iconColor,
+    required String label,
+  })  : _onPressed = onPressed,
+        _icon = icon,
+        _iconColor = iconColor,
+        _label = label;
+
+  final void Function()? _onPressed;
+  final IconData _icon;
+  final String _label;
+  final Color? _iconColor;
+
+  @override
+  Widget build(BuildContext context) {
+    Color mainColor = _iconColor ??
+        (_onPressed == null
+            ? Theme.of(context).colorScheme.onBackground.withOpacity(0.2)
+            : Theme.of(context).colorScheme.onBackground);
+    return Column(
+      children: [
+        OutlinedButton(
+          style: OutlinedButton.styleFrom(
+            padding: const EdgeInsets.all(24),
+            side: BorderSide(
+              color: mainColor,
+            ),
+          ),
+          onPressed: _onPressed,
+          child: Icon(
+            _icon,
+            color: mainColor,
+          ),
+        ),
+        const SizedBox(height: 16),
+        Text(
+          _label,
+          style: TextStyle(
+            color: mainColor,
+          ),
+        ),
+      ],
+    );
+  }
+}
